+++ backend-rust/src/graphql_api.rs
@@ -24,6 +24,7 @@ macro_rules! todo_api {
 pub(crate) use todo_api;
 
 use crate::{
+    connection::ConnectionQuery,
     scalar_types::{BlockHeight, DateTime, TimeSpan, UnsignedLong},
     transaction_event::smart_contracts::InvalidContractVersionError,
 };
@@ -414,68 +415,6 @@ pub fn get_pool<'a>(ctx: &Context<'a>) -> ApiResult<&'a PgPool> {
 /// Get service configuration from the context.
 pub fn get_config<'a>(ctx: &Context<'a>) -> ApiResult<&'a ApiServiceConfig> {
     ctx.data::<ApiServiceConfig>().map_err(ApiError::NoServiceConfig)
-}
-
-trait ConnectionCursor {
-    const MIN: Self;
-    const MAX: Self;
-}
-impl ConnectionCursor for i64 {
-    const MAX: i64 = i64::MAX;
-    const MIN: i64 = i64::MIN;
-}
-impl ConnectionCursor for usize {
-    const MAX: usize = usize::MAX;
-    const MIN: usize = usize::MIN;
-}
-
-struct ConnectionQuery<A> {
-    from:  A,
-    to:    A,
-    limit: i64,
-    // If the `last` elements are requested instead of the `first` elements
-    // (indicated by the `last` key being set when creating a new `ConnectionQuery`),
-    // the edges/nodes should be ordered in reverse (DESC) order before applying the range.
-    // This allows the range from `from` to `to` to be applied starting from the last element.
-    desc:  bool,
-}
-impl<A> ConnectionQuery<A> {
-    fn new<E>(
-        first: Option<u64>,
-        after: Option<String>,
-        last: Option<u64>,
-        before: Option<String>,
-        connection_limit: u64,
-    ) -> ApiResult<Self>
-    where
-        A: std::str::FromStr<Err = E> + ConnectionCursor,
-        E: Into<ApiError>, {
-        if first.is_some() && last.is_some() {
-            return Err(ApiError::QueryConnectionFirstLast);
-        }
-
-        let from = if let Some(a) = after {
-            a.parse::<A>().map_err(|e| e.into())?
-        } else {
-            A::MIN
-        };
-
-        let to = if let Some(b) = before {
-            b.parse::<A>().map_err(|e| e.into())?
-        } else {
-            A::MAX
-        };
-
-        let limit =
-            first.or(last).map_or(connection_limit, |limit| connection_limit.min(limit)) as i64;
-
-        Ok(Self {
-            from,
-            to,
-            limit,
-            desc: last.is_some(),
-        })
-    }
 }
 
 #[derive(Default)]
